+++ electron/main.ts
@@ -9,7 +9,8 @@ import {
   dialog,
   ipcMain,
   powerSaveBlocker,
-  protocol
+  protocol,
+  MenuItem
 } from 'electron'
 import {
   cpus,
@@ -53,11 +54,14 @@ import {
   iconDark,
   iconLight,
   installed,
+  kofiPage,
   legendaryBin,
   loginUrl,
+  patreonPage,
   sidInfoUrl,
   supportURL,
-  weblateUrl
+  weblateUrl,
+  wikiLink
 } from './constants'
 import { handleProtocol } from './protocol'
 import { listenStdout } from './logger'
@@ -163,7 +167,29 @@ async function createWindow(): Promise<BrowserWindow> {
       return handleExit()
     })
     mainWindow.loadURL(`file://${path.join(__dirname, '../build/index.html')}`)
-    mainWindow.setMenu(null)
+
+    const menu = new Menu()
+    menu.append(new MenuItem({
+      label: 'Help',
+      submenu: [
+        {
+          label: 'Reload',
+          accelerator: process.platform === 'darwin' ? 'Cmd+R' : 'Ctrl+R',
+          click: () => { mainWindow.reload() }
+        },
+        {
+          label: 'Debug',
+          accelerator: process.platform === 'darwin' ? 'Alt+Cmd+I' : 'Ctrl+Shift+I',
+          click: () => { mainWindow.webContents.openDevTools() }
+        },
+        {
+          label: 'Quit',
+          accelerator: process.platform === 'darwin' ? 'Cmd+Q' : 'Ctrl+Q',
+          click: () => { handleExit() }
+        }
+      ]
+    }))
+    mainWindow.setMenu(menu)
 
     return mainWindow
   }
@@ -214,7 +240,7 @@ const contextMenu = () => {
       label: i18next.t('tray.support', 'Support Us')
     },
     {
-      accelerator: 'ctrl + R',
+      accelerator: process.platform === 'darwin' ? 'Cmd+R' : 'Ctrl+R',
       click: function () {
         mainWindow.reload()
       },
@@ -224,7 +250,8 @@ const contextMenu = () => {
       click: function () {
         handleExit()
       },
-      label: i18next.t('tray.quit', 'Quit')
+      label: i18next.t('tray.quit', 'Quit'),
+      accelerator: process.platform === 'darwin' ? 'Cmd+Q' : 'Ctrl+Q'
     }
   ])
 }
@@ -261,7 +288,7 @@ if (!gotTheLock) {
       debug: false,
       fallbackLng: 'en',
       lng: language,
-      supportedLngs: ['ca', 'cs', 'de', 'el', 'en', 'es', 'fr', 'hr', 'hu', 'ja', 'ko', 'it', 'ml', 'nl', 'pl', 'pt', 'pt_BR', 'ru', 'sv', 'ta', 'tr', 'zh_Hans', 'zh_Hant']
+      supportedLngs: ['bg', 'ca', 'cs', 'de', 'el', 'en', 'es','fa', 'fi', 'fr', 'hr', 'hu', 'ja', 'ko', 'id', 'it', 'ml', 'nl', 'pl', 'pt', 'pt_BR', 'ru', 'sv', 'ta', 'tr', 'zh_Hans', 'zh_Hant']
 
     })
 
@@ -386,6 +413,9 @@ ipcMain.on('openWeblate', () => openUrlOrFile(weblateUrl))
 ipcMain.on('showAboutWindow', () => showAboutWindow())
 ipcMain.on('openLoginPage', () => openUrlOrFile(loginUrl))
 ipcMain.on('openDiscordLink', () => openUrlOrFile(discordLink))
+ipcMain.on('openPatreonPage', () => openUrlOrFile(patreonPage))
+ipcMain.on('openKofiPage', () => openUrlOrFile(kofiPage))
+ipcMain.on('openWikiLink', () => openUrlOrFile(wikiLink))
 ipcMain.on('openSidInfoPage', () => openUrlOrFile(sidInfoUrl))
 ipcMain.on('updateHeroic', () => checkUpdates())
 
@@ -457,6 +487,19 @@ ipcMain.handle('updateAll', () => LegendaryLibrary.get().updateAllGames())
 ipcMain.handle('checkVersion', () => checkForUpdates())
 
 ipcMain.handle('getMaxCpus', () => cpus().length)
+ipcMain.handle('getLegendaryVersion', async() => {
+  const { altLegendaryBin } = (await GlobalConfig.get().getSettings())
+  try {
+    if (altLegendaryBin && !altLegendaryBin.includes('legendary')) {
+      return 'invalid'
+    }
+    const {stdout} = await execAsync(`${legendaryBin} --version`)
+    return stdout.split('legendary version')[1].replaceAll('"', '').replaceAll(', codename', '')
+  } catch (error) {
+    return 'invalid'
+  }
+})
+
 ipcMain.handle('getPlatform', () => process.platform)
 
 ipcMain.on('createNewWindow', (e, url) =>
@@ -474,11 +517,16 @@ ipcMain.handle('getGameInfo', async (event, game) => {
 })
 
 ipcMain.handle('getInstallInfo', async (event, game) => {
+  const online = await isOnline()
+  if (!online){
+    return {game: {}, metadata: {}}
+  }
   try {
     const info = await Game.get(game).getInstallInfo()
     return info
   } catch (error) {
     logError(error)
+    return {}
   }
 })
 
@@ -513,6 +561,15 @@ ipcMain.handle('requestSettings', async (event, appName) => {
   return await GameConfig.get(appName).getSettings()
 })
 
+ipcMain.on('toggleDXVK', (event, [winePrefix, action]) => {
+  if (!existsSync(winePrefix)){
+    return
+  }
+
+  DXVK.installRemove(winePrefix, 'dxvk', action)
+  DXVK.installRemove(winePrefix, 'vkd3d', action)
+})
+
 ipcMain.handle('writeConfig', (event, [appName, config]) => {
   if (appName === 'default') {
     GlobalConfig.get().config = config
@@ -544,18 +601,23 @@ type RecentGame = {
   title: string
 }
 
-ipcMain.handle('launch', async (event, game: string) => {
+type LaunchParams = {
+  appName: string
+  launchArguments: string
+}
+
+ipcMain.handle('launch', async (event, {appName, launchArguments}: LaunchParams) => {
   const recentGames = store.get('games.recent') as Array<RecentGame> || []
-  const appName = game.split(' ')[0]
-  const { title } = await Game.get(appName).getGameInfo()
+  const game = appName.split(' ')[0]
+  const { title } = await Game.get(game).getGameInfo()
   const MAX_RECENT_GAMES = GlobalConfig.get().config.maxRecentGames || 5
   const startPlayingDate = new Date()
 
-  if (!tsStore.has(appName)){
-    tsStore.set(`${appName}.firstPlayed`, startPlayingDate)
+  if (!tsStore.has(game)){
+    tsStore.set(`${game}.firstPlayed`, startPlayingDate)
   }
 
-  logInfo('launching', title, appName)
+  logInfo('launching', title, game)
 
   if (recentGames.length) {
     let updatedRecentGames = recentGames.filter(a => a.appName !== game)
@@ -572,19 +634,19 @@ ipcMain.handle('launch', async (event, game: string) => {
     updatedRecentGames.unshift({ appName: game, title })
     store.set('games.recent', updatedRecentGames)
   } else {
-    store.set('games.recent', [{ appName: game, title: title }])
+    store.set('games.recent', [{ game, title: title }])
   }
 
-  return Game.get(game).launch().then(({ stderr }) => {
+  return Game.get(appName).launch(launchArguments).then(({ stderr }) => {
     const finishedPlayingDate = new Date()
-    tsStore.set(`${appName}.lastPlayed`, finishedPlayingDate)
+    tsStore.set(`${game}.lastPlayed`, finishedPlayingDate)
     const sessionPlayingTime = (Number(finishedPlayingDate) - Number(startPlayingDate)) / 1000 / 60
-    const totalPlayedTime: number = tsStore.has(`${appName}.totalPlayed`) ? tsStore.get(`${appName}.totalPlayed`) as number + sessionPlayingTime : sessionPlayingTime
+    const totalPlayedTime: number = tsStore.has(`${game}.totalPlayed`) ? tsStore.get(`${game}.totalPlayed`) as number + sessionPlayingTime : sessionPlayingTime
     // I'll send the calculated time here because then the user can set it manually on the file if desired
-    tsStore.set(`${appName}.totalPlayed`, Math.floor(totalPlayedTime))
+    tsStore.set(`${game}.totalPlayed`, Math.floor(totalPlayedTime))
 
     writeFile(
-      `${heroicGamesConfigPath}${appName}-lastPlay.log`,
+      `${heroicGamesConfigPath}${game}-lastPlay.log`,
       stderr,
       () => 'done'
     )