+++ src/App.test.tsx
@@ -5,11 +5,14 @@ import {
 } from '@testing-library/react';
 
 import App from './App';
-jest.mock('electron-store', () => jest.fn(() => ''));
+jest.mock('electron-store', () => jest.fn(() => {
+  return {configStore: {get: () => ''}}
+}));
+jest.mock('src//assets/heroic-icon.png', () => jest.fn(() => ''))
 
-describe.skip('App', () => {
+describe('App', () => {
 
-  test('renders', () => {
+  test.skip('renders', () => {
     render(<React.Suspense fallback="App loaded"><App /></React.Suspense>);
   })