+++ src/screens/Library/components/GameCard/index.tsx
@@ -114,7 +114,7 @@ const GameCard = ({
 
         setProgress(progress)
       }
-    }, 500)
+    }, 1500)
     return () => clearInterval(progressInterval)
   }, [isInstalling, appName])
 
@@ -270,7 +270,7 @@ const GameCard = ({
     }
     if (isInstalled) {
       await handleGameStatus({ appName, status: 'playing' })
-      return await launch(appName, t, handleGameStatus)
+      return await launch({appName, t, handleGameStatus})
     }
     return
   }