+++ src/screens/Wiki/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Wiki() {
-  return (
-    <div>
-      <webview id="foo" src='https://github.com/Heroic-Games-Launcher/HeroicGamesLauncher/wiki' style={{width:'100vw', height:'100vh'}}></webview>
-    </div>
-  )
-}