+++ src/test_helpers/testTypes.ts
@@ -117,6 +117,7 @@ class TestType<Type> {
  */
 const test_game = new TestType<GameInfo>({
   app_name: 'game',
+  canRunOffline: true,
   art_cover: 'art_cover',
   art_logo: 'art_logo',
   art_square: 'art_square',
@@ -155,6 +156,7 @@ const test_game = new TestType<GameInfo>({
  */
 const test_plugin = new TestType<GameInfo>({
   app_name: 'plugin',
+  canRunOffline: true,
   art_cover: 'art_cover',
   art_logo: 'art_logo',
   art_square: 'art_square',
@@ -192,6 +194,7 @@ const test_plugin = new TestType<GameInfo>({
  * Can be used in tests to provide/manipulate the app settings.
  */
 const test_appsettings = new TestType<AppSettings>({
+  altLegendaryBin: '',
   addDesktopShortcuts: false,
   addStartMenuShortcuts: false,
   audioFix: false,